+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add plugin installer script
 - Allow userdata update
+- Add encoded userdata option - to support templates
 
 ## 0.1.1 (August 22, 2017)